Bump batch size from uint8_t to uint16_t (#797)

+++ src/state.h
@@ -71,7 +71,7 @@ struct state_conf {
 	int cooldown_secs;
 	// number of sending threads
 	uint8_t senders;
-	uint8_t batch;
+	uint16_t batch;
 	uint32_t pin_cores_len;
 	uint32_t *pin_cores;
 	// should use CLI provided randomization seed instead of generating
